backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
requests==2.31.0
python-docx==0.8.11
pypdf==3.17.1
python-dotenv==1.0.0
aiofiles==23.2.1
pydantic==2.5.0